# [core] Add a real life benchmark