--- a/packages/material-ui-benchmark/README.md
+++ b/packages/material-ui-benchmark/README.md
@@ -1,5 +1,7 @@
 # Material-UI Benchmarking
 
+## Synthetic benchmark
+
 ```sh
 yarn benchmark
 
@@ -17,4 +19,31 @@ EmotionButton x 68,917 ops/sec ±19.32% (73 runs sampled)
 ButtonBase cache x 34,618 ops/sec ±12.83% (68 runs sampled)
 ButtonBase ripple x 38,715 ops/sec ±13.70% (68 runs sampled)
 ButtonBase disableRipple x 46,165 ops/sec ±13.53% (66 runs sampled)
+```
+
+## Real life benchmark
+
+```sh
+yarn server
+
+bombardier \
+  -c 100 \
+  -l \
+  -d 60s \
+  -m GET \
+  '0.0.0.0:3001/'
+
+Statistics        Avg      Stdev        Max
+  Reqs/sec       442.47      55.44     547.63
+  Latency      225.64ms    17.11ms   471.31ms
+  Latency Distribution
+     50%   221.98ms
+     75%   230.69ms
+     90%   241.19ms
+     95%   247.87ms
+     99%   273.88ms
+  HTTP codes:
+    1xx - 0, 2xx - 26642, 3xx - 0, 4xx - 0, 5xx - 0
+    others - 0
+  Throughput:    11.61MB/s
 ```