The page: https://material-ui.com/page-layout-examples/pricing/

**Without cache**:
```
Statistics        Avg      Stdev        Max
  Reqs/sec       160.27      37.58     249.11
  Latency      621.82ms    63.78ms      1.24s
  Latency Distribution
     50%   614.04ms
     75%   635.72ms
     90%   677.14ms
     95%   749.80ms
     99%   827.46ms
  HTTP codes:
    1xx - 0, 2xx - 9697, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     4.20MB/s
```

**With cache**:
```sh
Statistics        Avg      Stdev        Max
  Reqs/sec       442.47      55.44     547.63
  Latency      225.64ms    17.11ms   471.31ms
  Latency Distribution
     50%   221.98ms
     75%   230.69ms
     90%   241.19ms
     95%   247.87ms
     99%   273.88ms
  HTTP codes:
    1xx - 0, 2xx - 26642, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    11.61MB/s
```

JSS caching provides a nice x2.7 boost 🚀.